requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
freetype-py==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
pycairo==1.23.0
python-barcode==0.14.0
reportlab==4.0.0
rlPyCairo==0.2.0
Werkzeug==2.3.4
